+++ internal/daemon/hub/src/server/mod/seeder/yaml_rule_test.go
@@ -1,4 +1,4 @@
-package seedconfig
+package seeder
 
 import (
 	"os"